feat: organize code
