+++ app/src/main/java/br/com/useblu/oceands/client/ui/buttons/ButtonsActivity.kt
@@ -14,7 +14,7 @@ import androidx.lifecycle.ViewModelProvider
 import br.com.useblu.oceands.client.R
 import br.com.useblu.oceands.client.databinding.ActivityButtonsBinding
 import br.com.useblu.oceands.components.compose.OceanButton
-import br.com.useblu.oceands.components.compose.OceanButtonStyle
+import br.com.useblu.oceands.ui.compose.OceanButtonStyle
 import com.savvyapps.togglebuttonlayout.Toggle
 import com.savvyapps.togglebuttonlayout.ToggleButtonLayout
