+++ ocean-components/src/main/java/br/com/useblu/oceands/components/compose/OceanTopBarInverse.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
 import androidx.core.graphics.drawable.toBitmap
 import br.com.useblu.oceands.R
-import br.com.useblu.oceands.ui.compose.FontFamilyHighlightExtraBold
 import br.com.useblu.oceands.ui.compose.OceanColors
+import br.com.useblu.oceands.ui.compose.OceanFontFamily
 import br.com.useblu.oceands.ui.compose.OceanFontSize
 import br.com.useblu.oceands.ui.compose.OceanSpacing
 
@@ -90,7 +90,7 @@ fun OceanTopBarInverse(
             Text(
                 text = title,
                 fontSize = OceanFontSize.sm(),
-                fontFamily = FontFamilyHighlightExtraBold,
+                fontFamily = OceanFontFamily.HighlightExtraBold,
                 color = OceanColors.brandPrimaryPure()
             )
         }
