+++ ocean-components/src/main/java/br/com/useblu/oceands/ui/compose/OceanColors.kt
@@ -18,11 +18,89 @@ object OceanColors {
     fun interfaceLightDown() = colorResource(id = R.color.ocean_color_interface_light_down)
 
     @Composable
+    fun interfaceLightDeep() = colorResource(id = R.color.ocean_color_interface_light_deep)
+
+    @Composable
+    fun interfaceDarkPure() = colorResource(id = R.color.ocean_color_interface_dark_pure)
+
+    @Composable
     fun interfaceDarkUp() = colorResource(id = R.color.ocean_color_interface_dark_up)
 
     @Composable
     fun interfaceDarkDown() = colorResource(id = R.color.ocean_color_interface_dark_down)
 
     @Composable
+    fun interfaceDarkDeep() = colorResource(id = R.color.ocean_color_interface_dark_deep)
+
+    @Composable
     fun brandPrimaryPure() = colorResource(id = R.color.ocean_color_brand_primary_pure)
+
+    @Composable
+    fun brandPrimaryUp() = colorResource(id = R.color.ocean_color_brand_primary_up)
+
+    @Composable
+    fun brandPrimaryDown() = colorResource(id = R.color.ocean_color_brand_primary_down)
+
+    @Composable
+    fun brandPrimaryDeep() = colorResource(id = R.color.ocean_color_brand_primary_deep)
+
+    @Composable
+    fun highlightPure() = colorResource(id = R.color.ocean_color_highlight_pure)
+
+    @Composable
+    fun highlightUp() = colorResource(id = R.color.ocean_color_highlight_up)
+
+    @Composable
+    fun highlightDown() = colorResource(id = R.color.ocean_color_highlight_down)
+
+    @Composable
+    fun highlightDeep() = colorResource(id = R.color.ocean_color_highlight_deep)
+
+    @Composable
+    fun complementaryPure() = colorResource(id = R.color.ocean_color_complementary_pure)
+
+    @Composable
+    fun complementaryUp() = colorResource(id = R.color.ocean_color_complementary_up)
+
+    @Composable
+    fun complementaryDown() = colorResource(id = R.color.ocean_color_complementary_down)
+
+    @Composable
+    fun complementaryDeep() = colorResource(id = R.color.ocean_color_complementary_deep)
+
+    @Composable
+    fun statusPositivePure() = colorResource(id = R.color.ocean_color_status_positive_pure)
+
+    @Composable
+    fun statusPositiveUp() = colorResource(id = R.color.ocean_color_status_positive_up)
+
+    @Composable
+    fun statusPositiveDown() = colorResource(id = R.color.ocean_color_status_positive_down)
+
+    @Composable
+    fun statusPositiveDeep() = colorResource(id = R.color.ocean_color_status_positive_deep)
+
+    @Composable
+    fun statusNegativeUp() = colorResource(id = R.color.ocean_color_status_negative_up)
+
+    @Composable
+    fun statusNegativeDown() = colorResource(id = R.color.ocean_color_status_negative_down)
+
+    @Composable
+    fun statusNegativeDeep() = colorResource(id = R.color.ocean_color_status_negative_deep)
+
+    @Composable
+    fun statusNegativePure() = colorResource(id = R.color.ocean_color_status_negative_pure)
+
+    @Composable
+    fun statusWarningPure() = colorResource(id = R.color.ocean_color_status_warning_pure)
+
+    @Composable
+    fun statusWarningUp() = colorResource(id = R.color.ocean_color_status_warning_up)
+
+    @Composable
+    fun statusWarningDown() = colorResource(id = R.color.ocean_color_status_warning_down)
+
+    @Composable
+    fun statusWarningDeep() = colorResource(id = R.color.ocean_color_status_warning_deep)
 }
