+++ ocean-components/src/main/java/br/com/useblu/oceands/ui/compose/OceanFontFamily.kt
@@ -4,22 +4,24 @@ import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
 import br.com.useblu.oceands.R
 
-val FontFamilyBaseBold = FontFamily(Font(R.font.font_family_base_bold))
+object OceanFontFamily {
+    val BaseBold = FontFamily(Font(R.font.font_family_base_bold))
 
-val FontFamilyBaseExtraBold = FontFamily(Font(R.font.font_family_base_extrabold))
+    val BaseExtraBold = FontFamily(Font(R.font.font_family_base_extrabold))
 
-val FontFamilyBaseLight = FontFamily(Font(R.font.font_family_base_light))
+    val BaseLight = FontFamily(Font(R.font.font_family_base_light))
 
-val FontFamilyBaseMedium = FontFamily(Font(R.font.font_family_base_medium))
+    val BaseMedium = FontFamily(Font(R.font.font_family_base_medium))
 
-val FontFamilyBaseRegular = FontFamily(Font(R.font.font_family_base_regular))
+    val BaseRegular = FontFamily(Font(R.font.font_family_base_regular))
 
-val FontFamilyHighlightBold = FontFamily(Font(R.font.font_family_highlight_bold))
+    val HighlightBold = FontFamily(Font(R.font.font_family_highlight_bold))
 
-val FontFamilyHighlightExtraBold = FontFamily(Font(R.font.font_family_highlight_extrabold))
+    val HighlightExtraBold = FontFamily(Font(R.font.font_family_highlight_extrabold))
 
-val FontFamilyHighlightLight = FontFamily(Font(R.font.font_family_highlight_light))
+    val HighlightLight = FontFamily(Font(R.font.font_family_highlight_light))
 
-val FontFamilyHighlightMedium = FontFamily(Font(R.font.font_family_highlight_medium))
+    val HighlightMedium = FontFamily(Font(R.font.font_family_highlight_medium))
 
-val FontFamilyHighlightRegular = FontFamily(Font(R.font.font_family_highlight_regular))
+    val HighlightRegular = FontFamily(Font(R.font.font_family_highlight_regular))
+}
